Fixed merge conflicts and formatted the server files.

+++ controllers/client.js
@@ -1,11 +1,15 @@
 const { getDb } = require('../db/connect');
 const ObjectId = require('mongodb').ObjectId;
+const Api400Error = require('../error-handling/api400Error');
+const Api404Error = require('../error-handling/api404Error');
+const Api500Error = require('../error-handling/api500Error');
+const AsyncWrapper = require('../error-handling/AsyncWrapper');
 
 async function getDbCollection() {
   return getDb().db().collection('clients');
 }
 
-async function findAll(req, res) {
+const findAll = AsyncWrapper.wrapAsync(async (req, res, next) => {
   const collection = await getDbCollection();
   /*
     #swagger.tags = ['Clients']
@@ -19,47 +23,45 @@ async function findAll(req, res) {
     res.setHeader('Content-type', 'application/json');
     return res.status(200).json(clients);
   } catch (error) {
-    // Throwing custom errors.
-    throw new Api500Error(
-      "Retrieving clients failed",
-      undefined,
-      `Failed to retrieve clients, ${error.message}`, 
-    )
-    //console.log(error);
-    //return res.status(500).json("Oops! Something went wrong");
+    next(error);
   }
-}
+});
 
-async function findOne(req, res) {
-  const id = new ObjectId(req.params.id);
-  const collection = await getDbCollection();
+const findOne = AsyncWrapper.wrapAsync(async (req, res, next) => {
   /* 
     #swagger.tags = ['Clients']
     #swagger.description = 'Find a client by id'
   */
 
   try {
+    if (!ObjectId.isValid(req.params.id)) {
+      throw new Api400Error(
+        'Invalid Client ID Format',
+        undefined,
+        `The provided client ID ${req.params.id} is not a valid ID format.`
+      );
+    }
+    const id = new ObjectId(req.params.id);
+    const collection = await getDbCollection();
     const result = await collection.find({ _id: id });
     const clients = await result.toArray();
 
-    if (!clients.length) return res.status(404).json('Record not found');
+    if (!clients.length) {
+      throw new Api404Error(
+        'Client was not found',
+        undefined,
+        `Error occurred while searching client with ID ${req.params.id}.`
+      );
+    }
 
     res.setHeader('Content-type', 'application/json');
     return res.status(200).json(clients[0]);
   } catch (error) {
-    // Throwing custom errors.
-    throw new Api500Error(
-      "Retrieving client failed",
-      undefined,
-      `Failed to retrieve client with id ${req.params.id}, ${error.message}`,
-    );
-
-    //console.log(error);
-    //return res.status(500).json("Oops! Something went wrong");
+    next(error);
   }
-}
+});
 
-async function create(req, res) {
+const create = AsyncWrapper.wrapAsync(async (req, res, next) => {
   /*
    #swagger.tags = ['Clients']
    #swagger.description = 'Create a new client'
@@ -67,26 +69,8 @@ async function create(req, res) {
     in: 'body',
     description: 'Client information',
     required: true,
-    "schema": {
-      "type": "object",
-      "properties": {
-        "firstName": {
-          "example": "jacob"
-        },
-        "lastName": {
-          "example": "brown"
-        },
-        "email": {
-          "example": "jacob@email.com"
-        },
-        "phone": {
-          "example": "0756456352"
-        },
-        "dob": {
-          "example": "1990/02/05"
-        }
-      }
-    }
+    schema: { $ref: '#/definitions/Client' }
+   }
   */
   const { firstName, lastName, email, phone, dob } = req.body;
 
@@ -105,23 +89,19 @@ async function create(req, res) {
 
     if (result.acknowledged) {
       return res.status(201).json({ id: result.insertedId });
-    } else throw new Error('Failed to create new client');
+    } else {
+      throw new Api500Error(
+        'Client Creation Failed',
+        undefined,
+        'Failed to create a new client.'
+      );
+    }
   } catch (error) {
-
-    // Throwing custom errors.
-    throw new Api500Error(
-      "Client Creation Failed",
-      undefined,
-      `Error occurred while creating a new client, ${error.message}`,
-    );
-    //console.log(error);
-    //return res.status(500).json("Oops! Something went wrong");
+    next(error);
   }
-}
+});
 
-async function update(req, res) {
-  const id = new ObjectId(req.params.id);
-  const collection = await getDbCollection();
+const update = AsyncWrapper.wrapAsync(async (req, res, next) => {
   /*
    #swagger.tags = ['Clients']
    #swagger.description = 'Update a client by id'
@@ -133,19 +113,19 @@ async function update(req, res) {
       "type": "object",
       "properties": {
         "firstName": {
-          "example": "peter"
+          "example": "Peter"
         },
         "lastName": {
-          "example": "hamming"
+          "example": "Hamming"
         },
         "email": {
           "example": "peter@email.com"
         },
         "phone": {
-          "example": "0345243857"
+          "example": "340-524-3857"
         },
         "dob": {
-          "example": "1998/07/02"
+          "example": "07/02/1998"
         }
       }
     }
@@ -155,59 +135,69 @@ async function update(req, res) {
     $set: req.body,
   };
   try {
+    if (!ObjectId.isValid(req.params.id)) {
+      throw new Api400Error(
+        'Invalid Client ID Format',
+        undefined,
+        `The provided client ID ${req.params.id} is not a valid ID format.`
+      );
+    }
+    const id = new ObjectId(req.params.id);
+    const collection = await getDbCollection();
     const result = await collection.updateOne({ _id: id }, updateFilter);
 
     if (result.matchedCount === 0) {
-      return res.status(404).json('Record not found');
+      throw new Api404Error(
+        'Client was not found',
+        undefined,
+        `Error occurred while searching client with ID ${req.params.id}`
+      );
     }
 
     if (result.acknowledged) {
-      return res.sendStatus(204);
-    } else throw new Error("failed to update client");
-  } catch (error) 
-  {
-    // Throwing custom errors.
-    throw new Api500Error(
-      "Client Update Failed",
-      undefined,
-      `Error occurred while updating client with id ${req.params.id}, ${error.message}`,
-    );
-
-    //console.log(error);
-    //return res.status(500).json("Oops! Something went wrong");
+      res.status(200).json({ message: 'Client updated successfully' });
+    } else {
+      throw new Api500Error(
+        'Client Update Failed',
+        undefined,
+        `Error occurred while updating client with ID ${req.params.id}`
+      );
+    }
+  } catch (error) {
+    next(error);
   }
-}
+});
 
-async function remove(req, res) {
-  const id = new ObjectId(req.params.id);
-  const collection = await getDbCollection();
+const remove = AsyncWrapper.wrapAsync(async (req, res, next) => {
   /*
    #swagger.tags = ['Clients']
    #swagger.description = 'Remove a client by id'
   */
 
   try {
-    const result = await collection.deleteOne({ _id: id });
-
-    if (result.acknowledged && result.deletedCount == 0) {
-      return res.status(404).json('Record not found');
+    if (!ObjectId.isValid(req.params.id)) {
+      throw new Api400Error(
+        'Invalid Client ID Format',
+        undefined,
+        `The provided client ID ${req.params.id} is not a valid ID format.`
+      );
     }
+    const id = new ObjectId(req.params.id);
+    const collection = await getDbCollection();
+    const result = await collection.deleteOne({ _id: id });
 
     if (result.deletedCount > 0) {
       return res.status(200).json('Client was removed successfully');
-    } else throw new Error('Failed to delete client');
+    } else {
+      throw new Api404Error(
+        'Client was not found',
+        undefined,
+        `Error occurred while deleting client with ID ${req.params.id}`
+      );
+    }
   } catch (error) {
-    // Throwing custom errors.
-
-    throw new Api500Error(
-      "Client Deletion Failed",
-      undefined,
-      `Error occurred while deleting client with id ${req.params.id}, ${error.message}`,
-    );
-
-    //console.log(error);
-    //return res.status(500).json("Oops! Something went wrong");
+    next(error);
   }
-}
+});
 
 module.exports = { findAll, findOne, create, update, remove };
